+++ nmoo/evaluators/evaluation_penalty_evaluator.py
@@ -1,5 +1,5 @@
 """
-A custom evaluator that modifies the percieved number of evaluations
+A custom evaluator that modifies the perceived number of evaluations
 (`n_eval`).
 """
 __docformat__ = "google"
@@ -9,12 +9,12 @@ from pymoo.core.evaluator import Evaluator
 
 class EvaluationPenaltyEvaluator(Evaluator):
     """
-    A custom evaluator that modifies the percieved number of evaluations
+    A custom evaluator that miltiplies the perceived number of evaluations
     (`n_eval`).
 
-    For example, an algorithm using the following evaluator
+    For example, an pymoo algorithm using the following evaluator
 
-        EvaluationPenaltyEvaluator("times", 5)
+        EvaluationPenaltyEvaluator(5)
 
     will percieve every evaluation of the underlying problem as `5`
     evaluations. In other words, if the algorithm's maximum evaluation count is
@@ -22,25 +22,16 @@ class EvaluationPenaltyEvaluator(Evaluator):
     5` values.
     """
 
-    PENALTY_TYPES = ["times"]
+    _multiplier: int
 
-    _coefficient: int
-    _penalty_type: str
-
-    def __init__(self, penalty_type: str = "times", coefficient: int = 5):
+    def __init__(self, multiplier: int = 5):
         """
         Args:
-            penalty_type (str): the type of evaluation penalty to be applied.
-                Currently, must be `times`.
-            coefficient (int): coefficient pertaining to the penalty being
+            multiplier (int): multiplier pertaining to the penalty being
                 applied.
         """
         super().__init__()
-        self._coefficient = coefficient
-
-        if penalty_type not in EvaluationPenaltyEvaluator.PENALTY_TYPES:
-            raise ValueError(f"Unknown penalty type '{penalty_type}'")
-        self._penalty_type = penalty_type
+        self._multiplier = multiplier
 
     def eval(self, *args, **kwargs):  # pylint: disable=signature-differs
         """
@@ -48,6 +39,5 @@ class EvaluationPenaltyEvaluator(Evaluator):
         parameters.
         """
         result = super().eval(*args, **kwargs)
-        if self._penalty_type == "times":
-            self.n_eval += (self._coefficient - 1) * result.shape[0]
+        self.n_eval += (self._multiplier - 1) * result.shape[0]
         return result